refactor: removes print of "Summary:" from console output (#151)
<!-- 
Thanks for contributing to 2ms by offering a pull request.
-->

**Proposed Changes**
- removes summary from console output

We take the output from 2ms and create a file with it, and we run a
"validate sarif format" over it, and the console output has the
"Summary:" on top of it all, so I'm suggesting removing it

I submit this contribution under the Apache-2.0 license.

---------

Co-authored-by: Baruch Odem <baruch.odem@checkmarx.com>