fix: report directory is not exist (#207)
Closes #200 

I would usually start a discussion about a change before starting, but
since this was just a tiny fix I skipped it this time. Hope that's Ok.

This adds a path existence check to the path(s) passed in and ensures
the directory exists.

I submit this contribution under the Apache-2.0 license.

Co-authored-by: Baruch Odem (Rothkoff) <baruch.odem@checkmarx.com>