ci: update github actions versions (#213)
This pull request updates the versions of GitHub Actions used in the
workflow files. Specifically, it updates the `actions/checkout` action
from v3 to v4, and the `actions/setup-go` action from v4 to v5. These
updates ensure that the workflow files are using the latest versions of
the actions and take advantage of any bug fixes or new features
introduced in the newer versions.