feat: ast 40209 added validation status to sarif report (#238)
**Proposed Changes**

Added the property bag to Sarif Results ("properties"), in accordance
with SARIF guidelines.
Added validationStatus to properties of each result in SARIF, in line
with what already happens in JSON.

<!--
Please describe the big picture of your changes here. If it fixes a bug
or resolves a feature request, be sure to link to that issue.
-->

**Checklist**

- [ ] I covered my changes with tests (not applicable).
- [ ] I Updated the documentation that is affected by my changes (not
applicable):
  - [ ] Change in the CLI arguments
  - [ ] Change in the configuration file

I submit this contribution under the Apache-2.0 license.