fix: ast-68111 fix failing scan rule validation (#251)
<!--
Thanks for contributing to 2ms by offering a pull request.
-->

Closes #

**Proposed Changes**

<!--
There were issues with the validation of one rule, plaid-client-id.
Sometimes, auto-generated secrets during validation are bellow the
entropy threshold of the secret, causing the 2ms scan to fail.

In order to fix this the rule is now defined in 2ms and loaded from this
code, with the only difference being that the entropy value was changed
from 3.5 to 3.0.

vault-service-token was also added to 2ms because newer versions of
gitleaks have the same problem for that rule.

This is just a quick fix. The idea is move to a similar solution to
gitleaks, with rules being loaded from a config file instead of loaded
every scan.

-->

**Checklist**

- [ ] I covered my changes with tests.
- [ ] I Updated the documentation that is affected by my changes:
  - [ ] Change in the CLI arguments
  - [ ] Change in the configuration file

I submit this contribution under the Apache-2.0 license.