fix: thread exhaustion error (#297)
<!--
Thanks for contributing to 2ms by offering a pull request.
-->

**Proposed Changes**

The binary as is can take the process to resource exhaustion due to
going over the limit of 10k goroutines.

This is a quickfix/hotfix and in no case sustainable for the future, we
need indeed to refactor our current implementation but it should solve
the issue for now.

<!--
Please describe the big picture of your changes here. If it fixes a bug
or resolves a feature request, be sure to link to that issue.
-->

**Checklist**

- [ ] I covered my changes with tests.
- [ ] I Updated the documentation that is affected by my changes:
  - [ ] Change in the CLI arguments
  - [ ] Change in the configuration file

I submit this contribution under the Apache-2.0 license.