chore: make sure we have golangci-lint and every linter warning fixed (#305)
<!--
Thanks for contributing to 2ms by offering a pull request.
-->

**Proposed Changes**

This introduces a golangci-lint configuration, its warnings fixed, a
complete Makefile and pre-commit configuration
<!--
Please describe the big picture of your changes here. If it fixes a bug
or resolves a feature request, be sure to link to that issue.
-->

**Checklist**

- [ ] I covered my changes with tests.
- [ ] I Updated the documentation that is affected by my changes:
  - [ ] Change in the CLI arguments
  - [ ] Change in the configuration file

I submit this contribution under the Apache-2.0 license.