fix!: generate id vulnerability (#312)
<!--
Thanks for contributing to 2ms by offering a pull request.
-->

Closes #

**Proposed Changes**
Breaking change that fixes vulnerability to generate the result ID, this
is a breaking change because:
- IDs for the secrets are going to change
- If you were relying on the IDs for something as sorting the list of
results, that will also be impacted.

<!--
Please describe the big picture of your changes here. If it fixes a bug
or resolves a feature request, be sure to link to that issue.
-->

**Checklist**

- [ ] I covered my changes with tests.
- [ ] I Updated the documentation that is affected by my changes:
  - [ ] Change in the CLI arguments
  - [ ] Change in the configuration file

I submit this contribution under the Apache-2.0 license.

---------

Co-authored-by: Rogério Dalot <206845486+cx-rogerio-dalot-x@users.noreply.github.com>
Co-authored-by: Rui Oliveira <139987905+cx-rui-oliveira@users.noreply.github.com>