fix: introduce object pooling to process git diff changes (#320)
<!--
Thanks for contributing to 2ms by offering a pull request.
-->

Closes #

**Proposed Changes**
Adds object pooling for the git scan to improve memory usage and avoid
string builders to grow without bounds.
<!--
Please describe the big picture of your changes here. If it fixes a bug
or resolves a feature request, be sure to link to that issue.
-->

**Checklist**

- [X] I covered my changes with tests.
- [ ] I Updated the documentation that is affected by my changes:
  - [ ] Change in the CLI arguments
  - [ ] Change in the configuration file

I submit this contribution under the Apache-2.0 license.

---------

Co-authored-by: Rogério Dalot <206845486+cx-rogerio-dalot-x@users.noreply.github.com>