feat: upgrade gitleaks (#316)
<!--
Thanks for contributing to 2ms by offering a pull request.
-->

Closes #

**Proposed Changes**

- Upgrades gitleaks to the latest version and includes some of their
rules.
- Some of the rules were brought to our side since the gitleaks
validation logic is flaky and it fails our tests.
- Upgrades version to v4 that we forgot to upgrade when we bumped it on
the previous release.
- Adds command `make build-local` to makefile to build the binary
locally
- Gets rid of legacy script to lint the repo for print functions and
gets replaced by `forbidigo` linter in `golangci-lint`

<!--
Please describe the big picture of your changes here. If it fixes a bug
or resolves a feature request, be sure to link to that issue.
-->

**Checklist**

- [ ] I covered my changes with tests.
- [X] I Updated the documentation that is affected by my changes:
  - [ ] Change in the CLI arguments
  - [ ] Change in the configuration file

I submit this contribution under the Apache-2.0 license.

---------

Co-authored-by: Rogério Dalot <206845486+cx-rogerio-dalot-x@users.noreply.github.com>