feat: introduces workerpool (#306)
<!--
Thanks for contributing to 2ms by offering a pull request.
-->

Closes #

**Proposed Changes**

- It introduces a workerpool to optimize the detection which wraps
around [pond](https://github.com/alitto/pond)
- Initializes with a default number of workers calculated by number of
CPUs x2
- Reached that conclusion with the benchmark, see my comment below with
the stats.
- This workerpool can be used for other purposes in next steps of the
pipeline (in future PRs)
- On a future PR we won't wait for the detection to end.
- Fixed some linter issues
<!--
Please describe the big picture of your changes here. If it fixes a bug
or resolves a feature request, be sure to link to that issue.
-->

**Checklist**

- [X] I covered my changes with tests.
- [ ] I Updated the documentation that is affected by my changes:
  - [ ] Change in the CLI arguments
  - [ ] Change in the configuration file

I submit this contribution under the Apache-2.0 license.

---------

Co-authored-by: Rogério Dalot <206845486+cx-rogerio-dalot-x@users.noreply.github.com>