feat: support custom rules in CLI: target feature branch  (#347)
<!--
Thanks for contributing to 2ms by offering a pull request.
-->

Closes #

**Proposed Changes**
- Introduced flag --custom-rules-path which supports passing a file of
type JSON or YAML. This file has to be a list of ruledefine.Rule, and
can be used to define custom rules to be used in the scan

Custom Rules can be:
- Overrides - if a rule present in the file shares the same ruleId as a
default rule of 2ms, the rule present in the file will replace
(override) the default rule in the scan. If a rule is overridden, it
will simply take all fields from the rule as defined in the file. You
must include all fields that you want to be defined, otherwise they will
be nil/empty.
- New rules - if a rule does not share ruleId with a default rule, it
will be appended to the list of rules used in the scan.

- Rules can now be ignored by ruleId
- Custom rules work properly with --rule and --ignore-rule flags. Rules
can be selected/ignored by ruleId, ruleName and tag.

**Note:** Rules defined in the custom rules can find duplicate results
with each other or with default rules. 2ms doesn't match or correlate
results found by different rules, it will be the responsibility of the
user to avoid situations like these. **One exception** is
generic-api-key and overrides of generic-api-key, whose results will be
ignored if found by any other rule in the scan. This behavior already
happened in the default generic-api-key and was extended to overrides of
this rule.

**Checklist**

- [x] I covered my changes with tests.
- [x] I Updated the documentation that is affected by my changes:
  - [x] Change in the CLI arguments
  - [ ] Change in the configuration file

I submit this contribution under the Apache-2.0 license.