chore: update readme (#345)
<!--
Thanks for contributing to 2ms by offering a pull request.
-->

**Proposed Changes**

Update current README.md to make it more streamlined

<!--
Please describe the big picture of your changes here. If it fixes a bug
or resolves a feature request, be sure to link to that issue.
-->

I submit this contribution under the Apache-2.0 license.