feat: propagate resultId to SARIF (#359)
<!--
Thanks for contributing to 2ms by offering a pull request.
-->

Closes #

**Proposed Changes**
Have the  resultId in SARIF report.
I put the resultId in the field result.properties.resultId.

I was thinking of putting in the result.Guid field but the format does
not match the pattern of
"^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
in the SARIF schema.

https://schemastore.azurewebsites.net/schemas/json/sarif-2.1.0-rtm.5.json


<!--
Please describe the big picture of your changes here. If it fixes a bug
or resolves a feature request, be sure to link to that issue.
-->

**Checklist**

- [x] I covered my changes with tests.
- [ ] I Updated the documentation that is affected by my changes:
  - [ ] Change in the CLI arguments
  - [ ] Change in the configuration file

I submit this contribution under the Apache-2.0 license.